requirements.txt
fastapi
uvicorn
textblob
python-dotenv
requests